+++ app/Http/Controllers/ProjectsController.php
@@ -6,6 +6,7 @@ namespace App\Http\Controllers;
 
 use App\User;
 use App\Clients;
+use App\Invoices;
 use App\Projects;
 use App\Rules\Url;
 use Illuminate\Http\Request;
@@ -24,6 +25,13 @@ class ProjectsController extends Controller
     $this->page_settings['seltab'] = 'customers';
     $this->page_settings['seltab2'] = 'projects';
     $this->homeLink = '/projects';
+
+    $this->status = array( 
+      '1' => 'Open', 
+      '2' => 'Completed',
+      '3' => 'Dropped'
+    );
+
   }    
 
   public function index()
@@ -34,42 +42,76 @@ class ProjectsController extends Controller
 
       $active_status = (isset($_GET['active_status']) ? $_GET['active_status'] : 0);
 
-      if($active_status == 2){
+      if($active_status == 0){
         $dbtable = Projects::with('client:id,fname,lname')->where('is_categorized', 1)->orderBy('updated_at', 'DESC')->get();
       }else{
-        $dbtable = Projects::with('client:id,fname,lname')->where('is_deactivated', $active_status)->where('is_categorized', 1)->orderBy('updated_at', 'DESC')->get();
+        $dbtable = Projects::with('client:id,fname,lname')->where('status', $active_status)->where('is_categorized', 1)->orderBy('updated_at', 'DESC')->get();
       }
 
       return datatables()->of($dbtable)
         ->addColumn('action', function($data){
-      $button = '<div class="hover_buttons"><a href="/projects/view/'.$data->id.'" data-toggle="tooltip" data-placement="top" data-original-title="View" class="edit btn btn-outline-secondary btn-sm"><i class="fas fa-eye"></i></a>';
-      $button .= '<a href="/projects/edit/'.$data->id.'" data-toggle="tooltip" data-placement="top" data-original-title="Edit" class="edit btn btn-outline-secondary btn-sm edit-post"><i class="fas fa-edit"></i></a>';
-      $button .= '<a href="javascript:void(0);" id="delete-row" data-toggle="tooltip" data-placement="top" data-original-title="Delete" data-id="'.$data->id.'" class="delete btn-sm btn btn-outline-danger"><i class="fas fa-trash"></i></a></div>';
-      return $button;
+
+          $ditems = Invoices::where('projects_id', $data->id)->get();   
+          $sum = count($ditems);
+
+          $button = '<div class="hover_buttons"><a href="/projects/view/'.$data->id.'" data-toggle="tooltip" data-placement="top" data-original-title="View" class="edit btn btn-outline-secondary btn-sm"><i class="fas fa-eye"></i></a>';
+
+          if($data->status == 1){
+            if($sum != 0){
+              $button .= '<a href="javascript:void(0);" id="add-log-row" data-toggle="tooltip" data-placement="top" data-original-title="Set Status" data-id="'.$data->id.'"  class="edit btn btn-outline-secondary btn-sm"><i class="fas fa-history"></i></a>';
+            }
+
+            $button .= '<a href="/projects/edit/'.$data->id.'" data-toggle="tooltip" data-placement="top" data-original-title="Edit" class="edit btn btn-outline-secondary btn-sm edit-post"><i class="fas fa-edit"></i></a>';
+          }
+
+          if($sum == 0 && $data->status == 1){
+            $button .= '<a href="javascript:void(0);" id="delete-row" data-toggle="tooltip" data-placement="top" data-original-title="Delete" data-id="'.$data->id.'" class="delete btn-sm btn btn-outline-danger"><i class="fas fa-trash"></i></a></div>';
+          }
+
+          return $button;
       })
       ->addColumn('status',  function($data){
-        switch ($data->status) {
-          case 1:
-            $s = 'Open';
-            break;
-          case 2:
-            $s = 'Completed';
-            break;
-          case 3:
-            $s = 'Dropped';
-            break;
+        if($data->status){
+          $s = $this->status[$data->status];
         }
         return $s;
       })
+      ->addColumn('jobs',  function($data){
+        $ditems = Invoices::where('projects_id', $data->id)->get();   
+        $sum = count($ditems);
+        return $sum;
+      })
+      ->addColumn('url', function($data){
+        if(isset($data->url)) {
+          $url = $data->url;
+          if (strlen($url) >= 50) {
+            $n = shortenText($url, 50);
+          }
+          else {
+            $n = $url;
+          }
+
+          return '<a href="'.$url.'" target="blank">'.$n.'</a>';
+        }else{
+          return '';
+        }
+      })
+
       ->addColumn('created',  '{{ dateShortOnly($created_at) }}')
       ->addColumn('updated',  '{{ dateTimeFormat($updated_at) }}')
-      ->addColumn('checkbox', '<input type="checkbox" name="tbl_row_checkbox[]" class="tbl_row_checkbox" value="{{$id}}" />')
-        ->rawColumns(['checkbox','action','created','updated','status'])
+      ->addColumn('checkbox', function($data){
+        if($data->status == 1){
+          return '<input type="checkbox" name="tbl_row_checkbox[]" class="tbl_row_checkbox" value="'. $data->id .'" />';
+        }else{
+          return '&nbsp;';
+        }
+      })
+        ->rawColumns(['checkbox','action','status','url'])
       ->make(true);
       
     }
         
-    return view('projects.index', ['user' => $user, 'page_settings'=> $this->page_settings]);
+    return view('projects.index', ['user' => $user, 'page_settings'=> $this->page_settings, 'status' => $this->status]);;
 
   }
 
@@ -91,7 +133,6 @@ class ProjectsController extends Controller
       'url' => ['nullable', 'string', 'max:255', new Url],
       'description' => ['nullable'],
       'client_id' => ['required'],
-      'status' => ['required'],
     ]);
 
     $client_check = Clients::find($data['client_id']);
@@ -105,7 +146,7 @@ class ProjectsController extends Controller
       'url' => $data['url'],
       'description' => $data['description'],
       'client_id' => $data['client_id'],
-      'status' => $data['status'],
+      'status' => 1,
       'is_categorized' => 1, 
       'is_deactivated' => 0, 
       'updatedby_id' => $user->id,
@@ -124,18 +165,9 @@ class ProjectsController extends Controller
     $project = Projects::with('client')->find($id);
     $updater = User::find($project->updatedby_id);
 
-    switch ($project->status) {
-      case 1:
-        $s = 'Open';
-        break;
-      case 2:
-        $s = 'Completed';
-        break;
-      case 3:
-        $s = 'Dropped';
-        break;
+    if($project->status){
+      $s = $this->status[$project->status];
     }
-
 
 
     return view('projects.view', ['user' => $user, 'project' => $project, 'page_settings'=> $this->page_settings, 'updater' => $updater, 's'=> $s]);
@@ -171,7 +203,6 @@ class ProjectsController extends Controller
       'url' => ['nullable', 'string', 'max:255', new Url],
       'description' => ['nullable'],
       'client_id' => ['required'],
-      'status' => ['required'],
     ]);
 
 
@@ -186,7 +217,6 @@ class ProjectsController extends Controller
     $project->url = $data['url'];
     $project->description = $data['description'];
     $project->client_id = $data['client_id'];
-    $project->status = $data['status'];
     $project->updatedby_id = $user->id;
 
     $query = $project->update();
@@ -194,7 +224,6 @@ class ProjectsController extends Controller
     if($query){
       return notifyRedirect('/projects/view/'.$id, 'Updated project '. $project->name .' successfully', 'success');
     }
-
   }
 
   public function destroy($id)
@@ -203,7 +232,13 @@ class ProjectsController extends Controller
     $project = Projects::find($id);
     if($project){
       if(request()->ajax()){
-        $row = Projects::where('id',$id)->delete();
+        $ditems = Invoices::where('projects_id', $id)->get();   
+        $sum = count($ditems);
+
+        if($sum == 0){
+          $row = Projects::where('id',$id)->delete();
+        }
+        
         return Response::json($row);
       }else{
         return notifyRedirect($this->homeLink, 'Unauthorized to delete', 'danger');
@@ -215,30 +250,49 @@ class ProjectsController extends Controller
 
 
 
+
   public function status(Request $request)
   {
     if(request()->ajax()){
-      
-      $row_id_array = $request->input('id');
+      $req_id = $request->input('id');
       $form = $request->input('formData');
 
-      $count_updated = 0;
+      if(is_array($req_id)){
+        $count_updated = 0;
 
-      foreach ($row_id_array as $row) {
-          $row = Projects::find($row); 
+        foreach ($req_id as $row_id) {
+
+            $row = Projects::find($row_id); 
+            if($row){
+              $row->status = $form['status'];
+              $row->updatedby_id = $form['updatedby_id'];
+              $row->update();
+
+              $count_updated++;
+            }
+        }
+        return Response::json($count_updated);
+
+      }else{
+
+        $row = Projects::find($req_id); 
+        if($row){
           $row->status = $form['status'];
           $row->updatedby_id = $form['updatedby_id'];
-          $query = $row->update();
-          $count_updated++;
+          $row->update();
+
+          return Response::json(1);
+        }
       }
 
-      return Response::json($count_updated);
 
     }else{
-      return notifyRedirect($this->homeLink, 'Deletion action not permitted', 'danger');
+      return notifyRedirect($this->homeLink, 'Action not permitted', 'danger');
     }
-
   }
+
+
+
 
   public function invoiceautocomplete(Request $request)
   {
